rename : message/in, out -> message/input, output
