+++ src/main/kotlin/io/toasting/domain/message/applicatoin/input/CreateChatRoomInput.kt
@@ -1,4 +1,4 @@
-package io.toasting.domain.message.applicatoin.`in`
+package io.toasting.domain.message.applicatoin.input
 
 class CreateChatRoomInput(
     val partnerId: Long
