+++ src/main/kotlin/io/toasting/domain/message/applicatoin/output/CreateChatRoomOutput.kt
@@ -1,4 +1,4 @@
-package io.toasting.domain.message.applicatoin.out
+package io.toasting.domain.message.applicatoin.output
 
 import io.toasting.domain.message.entity.ChatRoom
 import java.time.LocalDateTime
