+++ src/main/kotlin/io/toasting/domain/message/applicatoin/output/GetMessageCountOutput.kt
@@ -1,4 +1,4 @@
-package io.toasting.domain.message.applicatoin.out
+package io.toasting.domain.message.applicatoin.output
 
 class GetMessageCountOutput(
     val count: Long
