+++ src/main/kotlin/io/toasting/domain/message/controller/request/SendMessageRequest.kt
@@ -1,6 +1,6 @@
 package io.toasting.domain.message.controller.request
 
-import io.toasting.domain.message.applicatoin.`in`.SendMessageInput
+import io.toasting.domain.message.applicatoin.input.SendMessageInput
 import jakarta.validation.constraints.NotBlank
 
 data class SendMessageRequest(
