+++ src/main/kotlin/io/toasting/domain/message/controller/response/CreateChatRoomResponse.kt
@@ -1,6 +1,6 @@
 package io.toasting.domain.message.controller.response
 
-import io.toasting.domain.message.applicatoin.out.CreateChatRoomOutput
+import io.toasting.domain.message.applicatoin.output.CreateChatRoomOutput
 import java.time.LocalDateTime
 
 class CreateChatRoomResponse(
