+++ src/test/kotlin/io/toasting/domain/message/application/MessageServiceTest.kt
@@ -15,8 +15,8 @@ import io.toasting.domain.member.entity.Member
 import io.toasting.domain.member.entity.MemberDetails
 import io.toasting.domain.member.repository.MemberRepository
 import io.toasting.domain.message.applicatoin.MessageService
-import io.toasting.domain.message.applicatoin.`in`.CreateChatRoomInput
-import io.toasting.domain.message.applicatoin.`in`.SendMessageInput
+import io.toasting.domain.message.applicatoin.input.CreateChatRoomInput
+import io.toasting.domain.message.applicatoin.input.SendMessageInput
 import io.toasting.domain.message.entity.ChatMember
 import io.toasting.domain.message.entity.ChatRoom
 import io.toasting.domain.message.entity.Message
